JAMES-3534 Move Identity to data-jmap
